Use fixed hash width for `ty_server` diagnostics
Summary
--

Fixes a snapshot test failure I saw in #19653 locally and in Windows CI by
padding the hex ID to 16 digits to match the regex in `filter_result_id`.

https://github.com/astral-sh/ruff/blob/78e5fe0a51fcbb1456d54c65b4632f60c268e389/crates/ty_server/tests/e2e/pull_diagnostics.rs#L380-L384

Test Plan
--

I applied this to the branch from #19653 locally and saw that the tests now
pass. I couldn't reproduce this failure directly on `main` or this branch,
though.