accept other snapshots
The ISC snapshots are definitely correct as they are the ones that indicated
there was a problem in ruff_db. I believe the JSON snapshot is correct too. The
input is this:

```py
import os
```

with a trailing newline. The old edit end location was reported as 2:1 instead
of 1:10. I think it's accurate to use 1:10 for this too.