Avoid underflow in default ranges before a BOM
Summary
--

This fixes a regression caused by the BOM handling in #19806. Most diagnostics
already account for the BOM in their ranges, but those that use
`TextRange::default` to mean the beginning of the file do not, causing an
underflow in `RenderableAnnotation::new` when subtracting the BOM-shifted
`snippet_start` from the annotation range.

I ran into this when trying to run benchmarks on CPython in preparation for
caching work. The file `cpython/Lib/test/bad_coding2.py` was causing a crash
because it had a default-range `I002` diagnostic, with a BOM.

https://github.com/astral-sh/ruff/blob/7cc3f1ebe9386e77e7009bc411fc6480d3851015/crates/ruff_linter/src/rules/isort/rules/add_required_imports.rs#L122-L126

The fix here is just to saturate to zero instead of panicking. I considered
adding a `TextRange::saturating_sub` method, but I wasn't sure it was worth it
for this one use. I'm happy to do that if preferred, though.

Saturating seemed easier than shifting the affected annotations over, but that
could be another solution.

Test Plan
--

A new `ruff_db` test that reproduced the issue and manual testing against the
CPython file mentioned above