Don't cache files with diagnostics
Summary
--

To take advantage of the new diagnostics, we need to update our caching model to
include all of the information supported by `ruff_db`'s diagnostic type. Instead
of trying to serialize all of this information, Micha suggested simply not
caching files with diagnostics, like we already do for files with syntax errors.
This PR is a prototype of that approach.

Test Plan
--

Existing tests, with their input updated not to include a diagnostic, plus a new
test showing that files with lint diagnostics are not cached.

Benchmarks
--

In addition to tests, we wanted to check that this doesn't degrade performance
too much. I posted part of this new analysis in
https://github.com/astral-sh/ruff/issues/18198#issuecomment-3175048672, but I'll
duplicate it here. In short, there's not much difference between `main` and this
branch for projects with few diagnostics (`home-assistant`, `airflow`), as
expected. And the difference for projects with many diagnostics (`cpython`)
isn't that big either. As discussed in #18198, printing the large number of
diagnostics starts to dominate the CPython timing anyway.

| Command                                                       | Mean [ms] | Min [ms] | Max [ms] |
|:--------------------------------------------------------------|----------:|---------:|---------:|
| `ruff check cpython --no-cache --isolated --exit-zero`        |     322.0 |    317.5 |    326.2 |
| `ruff check cpython --isolated --exit-zero`                   |     217.3 |    209.8 |    237.9 |
| `ruff check home-assistant --no-cache --isolated --exit-zero` |     279.5 |    277.0 |    283.6 |
| `ruff check home-assistant --isolated --exit-zero`            |      37.2 |     35.7 |     40.6 |
| `ruff check airflow --no-cache --isolated --exit-zero`        |     133.1 |    130.4 |    146.4 |
| `ruff check airflow --isolated --exit-zero`                   |      34.7 |     32.9 |     41.6 |
|:--------------------------------------------------------------|----------:|---------:|---------:|
| `ruff check cpython --no-cache --isolated --exit-zero`        |     330.1 |    324.5 |    333.6 |
| `ruff check cpython --isolated --exit-zero`                   |     309.2 |    306.1 |    314.7 |
| `ruff check home-assistant --no-cache --isolated --exit-zero` |     288.6 |    279.4 |    302.3 |
| `ruff check home-assistant --isolated --exit-zero`            |      39.8 |     36.9 |     42.4 |
| `ruff check airflow --no-cache --isolated --exit-zero`        |     134.5 |    131.3 |    140.6 |
| `ruff check airflow --isolated --exit-zero`                   |      39.1 |     37.2 |     44.3 |

I had Claude adapt one of the scripts from the hyperfine repo to make this plot,
so it's not quite perfect, but maybe it's still useful. The table is probably
more reliable for close comparisons. I'll put more details about the benchmarks
below for the sake of future reproducibility.

<plot>

<details>