remove LinterResult::has_no_syntax_errors flag and helper methods
in the previous commit, this check was replaced by `diagnostics.is_empty()`
because any syntax errors will already have been converted to diagnostics. here,
we revert the assertion in the benchmark to use only parse errors, as the
comment still said and inline the other local uses of the other helper methods.
`has_invalid_syntax` is still used in the LSP, where we don't have access to the
parsed module.