[`pyupgrade`] Add sub-diagnostics showing type variable definitions (`UP046`)
Summary
--

Now that we've handled caching for the new diagnostics, I went looking for
diagnostics that could use additional annotations or sub-diagnostics that
weren't already tracked in #17203. UP046 (and UP040 and UP047 if we like this
one) seemed like a reasonable candidate because we check for separate type
variable definitions in the same file before emitting a diagnostic. In this
case, I added an info-level sub-diagnostic pointing to this type variable
definition. I'm not actually sure if this is super helpful information, so I
wouldn't mind closing this if others don't think it's useful.

I also added a helper `DiagnosticGuard::info` method that I think will be useful
for adding this kind of sub-diagnostic in general.

I spent a while debugging why the snapshots were showing a longer path in the
sub-diagnostics before remembering that we overwrite the source file for the
main diagnostic here:

https://github.com/astral-sh/ruff/blob/79c949f0f75f6582782f3d5fddc65089789801d5/crates/ruff_linter/src/test.rs#L280-L288

As shown below, the sub-diagnostics look as expected in the CLI. I think we
_could_ delete this part of the test code, but it results in using the longer
test file names (like our new sub-diagnostics) in every snapshot. So
alternatively, we may want to apply the same transformation to the
sub-diagnostics (and secondary annotations) too.

Test Plan
--

Existing UP046 tests with newly expanded snapshots, as well as some manual
testing in the CLI: