Render concise snapshots in linter tests
Summary
--

This is a quick implementation of the idea in
https://github.com/astral-sh/ruff/pull/19900#issuecomment-3188463902. Especially
as we start using new diagnostic features, it seems like a good idea to keep an
eye on the `concise` output format as well as the `full`. This PR just appends a
second rendering of the diagnostics in our rule snapshots in the `concise`
format. The concise message is typically reused in the other output formats too,
so this gives us a bit of coverage for all the output formats.

I noticed that some of our concise messages are still quite long. That might be
another good class of diagnostic to use new features for. Some explanatory
pieces might fit better as sub-diagnostics.

Test Plan
--

Existing snapshots. Again, I clicked through these individually, but very, very
quickly