Solve unbound typevar to Unknown & UpperBound
Was just playing around with intersections. I think this is a thing that
makes sense and has also come up in real life.

It looks like it is an improvement on some test cases, mostly because ty
doesn't bind Self. Need to look into the duplicates and the scary cycle
error