Add a section on the new diagnostic format
Summary
--

I thought this might warrant a small blog-style writeup, especially since we
already got a question about it (#19966), but I'm happy to switch back to a
one-liner under `### Other changes` if preferred.

I'll copy whatever we add here to the release notes too.

Do we need a note at the top about the late addition?